requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
python-jose
passlib[bcrypt]
jinja2
cryptography
pytest
pytest-asyncio
pytest-mock
pytest-cov
httpx